Make a ROCM FA backend
