+++ vllm/attention/selector.py
@@ -1,4 +1,3 @@
-import os
 from functools import lru_cache
 from typing import Type
 
@@ -18,11 +17,11 @@ def get_attn_backend(dtype: torch.dtype) -> Type[AttentionBackend]:
         from vllm.attention.backends.flash_attn import (  # noqa: F401
             FlashAttentionBackend)
         return FlashAttentionBackend
-    elif _which_attn_to_use(dtype) == "TritonFlashAttention":
-        logger.info("Using TritonFlashAttention backend.")
-        from vllm.attention.backends.triton_flash_attn import (  # noqa: F401
-            TritonFlashAttentionBackend)
-        return TritonFlashAttentionBackend
+    elif _which_attn_to_use(dtype) == "ROCmFlashAttention":
+        logger.info("Using ROCmFlashAttention backend.")
+        from vllm.attention.backends.rocm_flash_attn import (  # noqa: F401
+            ROCmFlashAttentionBackend)
+        return ROCmFlashAttentionBackend
     else:
         logger.info("Using XFormers backend.")
         from vllm.attention.backends.xformers import (  # noqa: F401
@@ -34,14 +33,8 @@ def _which_attn_to_use(dtype: torch.dtype) -> str:
     """Returns which flash attention backend to use.
 
     Returns:
-        str: XFormers, FlashAttention, or TritonFlashAttention
+        str: XFormers, FlashAttention, or ROCmFlashAttention
     """
-
-    # NOTE: Allow for switching between Triton and FA
-    #       Defaulting to triton FA for AMD cards.
-    use_triton_flash_attn = (os.environ.get("VLLM_USE_TRITON_FLASH_ATTN",
-                                            "True").lower()
-                             in ("true", "1")) and is_hip()
 
     if not is_hip() and torch.cuda.get_device_capability()[0] < 8:
         # Volta and Turing NVIDIA GPUs.
@@ -60,8 +53,8 @@ def _which_attn_to_use(dtype: torch.dtype) -> str:
                     "torch.float16 or torch.bfloat16.")
         return "XFormers"
 
-    if not use_triton_flash_attn:
-        # Only test for flash_attn if we are using it.
+    if not is_hip():
+        # ROCm backend has its own check for this.
         try:
             import flash_attn  # noqa: F401
         except ImportError:
@@ -70,4 +63,4 @@ def _which_attn_to_use(dtype: torch.dtype) -> str:
                 "Please install it for better performance.")
             return "XFormers"
 
-    return "TritonFlashAttention" if use_triton_flash_attn else "FlashAttention"
+    return "ROCmFlashAttention" if is_hip() else "FlashAttention"
